FRBID_code/requirements.txt
pandas==0.24.1
numpy==1.14.6
tensorflow-gpu==1.9.0
imbalanced_learn==0.4.3
matplotlib==3.0.3
scipy==1.2.1
Keras==2.0.9
imblearn==0.0
Pillow==7.1.2
scikit_learn==0.23.1
